AssertSource refactoring
